# docs/03-architecture/api-spec.yaml
openapi: 3.1.0
info:
  title: Smart Cab API
  version: 0.1.0
  description: |
    Contract between the Flutter app, simulator and FastAPI backend.
    RULE: change this file first, regenerate the Dart client (`make api-client`), then implement.
    Phase 1 endpoints are complete. Phase 2+ endpoints are marked `x-phase` and may be refined before implementation.
    All times are ISO 8601 UTC. All requests except /auth/* need `Authorization: Bearer <access token>` and `X-Active-Role`.
    Errors use the `Error` schema. Pagination: `limit` (max 100) + `cursor`.
    The realtime surface (`/ws`) cannot be described in OpenAPI; its contract is
    `websocket-protocol.md`, referenced below under `x-websocket`, and the same
    change-the-contract-first rule applies to it.
...
servers:
  - url: http://localhost:8000/api/v1
    description: local dev
security:
  - bearerAuth: []
...
paths:
...
  /sos:
    post:
      tags: [employee, driver]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [lat, lng]
              properties:
                trip_id: {type: [string, 'null'], format: uuid}
                lat: {type: number}
                lng: {type: number}
      responses:
        '201': {description: Alert raised}
...
  /driver/issues:
    post:
      tags: [driver]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [type]
              properties:
                type: {type: string, enum: [breakdown, accident, traffic_block, rider_issue, other]}
                note: {type: string, maxLength: 300}
                trip_id: {type: [string, 'null'], format: uuid}
                lat: {type: number}
                lng: {type: number}
      responses:
        '201': {description: Alert raised}
...
  /alerts:
    get:
      tags: [alerts]
      parameters:
        - {name: status, in: query, schema: {type: string, enum: [open, acknowledged, resolved]}}
      responses:
        '200':
          description: Alerts
          content:
            application/json:
              schema: {type: array, items: {$ref: '#/components/schemas/Alert'}}
  /alerts/{alert_id}/{action}:
    parameters:
      - {name: alert_id, in: path, required: true, schema: {type: string, format: uuid}}
      - {name: action, in: path, required: true, schema: {type: string, enum: [acknowledge, resolve]}}
    post:
      tags: [alerts]
      responses:
        '200': {description: Updated}
...
  /simctl/clock:
    x-env: sim-only
    get:
      tags: [simctl]
      responses:
        '200':
          description: Current sim time
          content:
            application/json:
              schema: {$ref: '#/components/schemas/SimClock'}
    put:
      tags: [simctl]
      summary: Set or advance sim time (APP_ENV=sim only; 404 otherwise)
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/SimClock'}
      responses:
        '200': {description: Updated}
  /simctl/reset:
    x-env: sim-only
    post:
      tags: [simctl]
      summary: Truncate operator data and load scenario seed
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                scenario_yaml: {type: string}
      responses:
        '200':
          description: Seeded entity ids and test user tokens
          content:
            application/json:
              schema: {type: object, additionalProperties: true}
components:
  securitySchemes:
    bearerAuth: {type: http, scheme: bearer, bearerFormat: JWT}
...
  schemas:
...
    Alert:
      type: object
      properties:
        id: {type: string, format: uuid}
        type: {type: string, enum: [sos, driver_issue, request_near_expiry, vip_no_vehicle, failsafe, stale_vehicle, mode_prompt, system]}
        severity: {type: string, enum: [critical, warning, info]}
        status: {type: string, enum: [open, acknowledged, resolved]}
        created_at: {type: string, format: date-time}
        request_id: {type: [string, 'null'], format: uuid}
        trip_id: {type: [string, 'null'], format: uuid}
        vehicle_id: {type: [string, 'null'], format: uuid}
        data: {type: object, additionalProperties: true}
...
    SimClock:
      type: object
      properties:
        now: {type: string, format: date-time}
        advance_seconds: {type: integer, description: PUT only, alternative to now}